refactor(zim-to-website): move article processing out into its own file
