+++ zim-to-website/src/site-transforms.ts
@@ -1,0 +1,73 @@
+import {
+  readdirSync,
+  readFileSync,
+  writeFileSync,
+  lstatSync,
+  copyFileSync,
+  renameSync,
+  existsSync
+} from 'fs'
+import { join } from 'path'
+import Handlebars from 'handlebars'
+import { Options, Directories } from './zim-to-website'
+
+const indexRedirectFragment = readFileSync(
+  './src/index_redirect_fragment.handlebars'
+)
+
+export const copyImageAssetsIntoWiki = async (
+  assetsDir: string,
+  imagesFolder: string
+) => {
+  const imagesFiles = readdirSync(assetsDir)
+
+  for (const imageFile of imagesFiles) {
+    const filepath = join(assetsDir, imageFile)
+    const info = lstatSync(filepath)
+
+    if (!info.isFile()) {
+      return
+    }
+
+    const imagesFolderPath = join(imagesFolder, imageFile)
+    copyFileSync(filepath, imagesFolderPath)
+  }
+}
+
+export const moveArticleFolderToWiki = (
+  articleFolder: string,
+  wikiFolder: string
+) => {
+  if (existsSync(wikiFolder)) {
+    return
+  }
+
+  renameSync(articleFolder, wikiFolder)
+}
+
+export const insertIndexRedirect = (options: Options) => {
+  const template = Handlebars.compile(indexRedirectFragment.toString())
+
+  const indexPath = join(options.unpackedZimDir, 'index.html')
+  writeFileSync(
+    indexPath,
+    template({
+      MAIN_PAGE: options.mainPage
+    })
+  )
+}
+
+export const resolveDirectories = (options: Options) => {
+  const articleFolder = join(options.unpackedZimDir, 'A')
+  const imagesFolder = join(options.unpackedZimDir, 'I', 'm')
+  const wikiFolder = join(options.unpackedZimDir, 'wiki')
+
+  const directories: Directories = {
+    unpackedZimDir: options.unpackedZimDir,
+    articleFolder,
+    imagesFolder,
+    wikiFolder
+  }
+
+  return directories
+}
