+++ zim-to-website/src/utils/walk-files.ts
